Major cleanup separated callbacks into its own file need to sort out what libraries are actually used.
