+++ INPUT DATA/SOLAR_REPRESENTATIVE_YEAR_60_MINUTES.xlsx
