+++ config.py
@@ -1,0 +1,5 @@
+## file to define global variables ##
+
+Weekly_Interval_Data = []
+Daily_Interval_Data = []
+Monthly_Sum = []
