+++ myproject/wsgi.py
@@ -1,0 +1,3 @@
+from .app import server as application 
+#IMPORT USER DEFINED GLOBAL VARIABLES 
+import config
